requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
voyageai==0.2.4
Pillow==10.0.1
numpy==1.24.3
python-dotenv==1.0.0
